startup-funding-tracker/requirements.txt
streamlit
pandas
plotly
numpy
python-dotenv
sqlite-utils
